refactor(mastracode): move web router page components into pages folder
All route-level page components (SignInPage, NewPage, ThreadPage,
OverviewPage, BoardPage, MetricsPage, AuditPage) lived scattered across
domains/auth, domains/chat, and domains/factory while CreateFactoryPage
and OnboardingPage were already in pages/. Consolidate every page the
router mounts into src/web/ui/pages so the route table maps 1:1 to the
folder, and drop the now-stale page re-exports from the domain indexes.

Chat (pathless layout) and RootGuards stay in their domains since they
are not pages.
