+++ mastracode/web/src/web/ui/__tests__/message-rendering.msw.test.tsx
@@ -12,8 +12,8 @@ import { server } from '../../../../e2e/web-ui/msw-server';
 import { renderWithProviders, TEST_BASE_URL } from '../../../../e2e/web-ui/render';
 import { loginUrl, logoutUrl } from '../domains/auth';
 import Chat from '../domains/chat/Chat';
-import { NewPage } from '../domains/chat/NewPage';
-import { ThreadPage } from '../domains/chat/ThreadPage';
+import { NewPage } from '../pages/NewPage';
+import { ThreadPage } from '../pages/ThreadPage';
 import type { Factory } from '../domains/workspaces';
 import { ActiveFactoryProvider } from '../domains/workspaces/context/ActiveFactoryProvider';
 import { CreateFactoryPage } from '../pages/CreateFactoryPage';
