+++ mastracode/web/src/web/ui/domains/auth/index.ts
@@ -1,4 +1,3 @@
-export { safeReturnTo, SignInPage } from './components/SignInPage';
 export { useFactoryAuth } from '../../../../shared/hooks/useFactoryAuth';
 export {
   fetchAuthState,
