+++ mastracode/web/src/web/ui/domains/factory/index.ts
@@ -1,6 +1,3 @@
-export { AuditPage } from './AuditPage';
-export { BoardPage, ReviewBoardPage, WorkBoardPage } from './BoardPage';
-export { MetricsPage } from './MetricsPage';
 export { FactorySection } from './components/FactorySection';
 export * from '../../../../shared/hooks/useFactoryData';
 export * from '../../../../shared/hooks/useLinearData';
