+++ mastracode/web/src/web/ui/pages/MetricsPage.tsx
@@ -5,16 +5,16 @@ import { Notice } from '@mastra/playground-ui/components/Notice';
 import { Txt } from '@mastra/playground-ui/components/Txt';
 import { useState } from 'react';
 
-import { useApiConfig } from '../../../../shared/api/config';
-import { useFactoryMetrics } from '../../../../shared/hooks/useFactoryMetrics';
-import { useWorkspaceActivity } from '../../../../shared/hooks/useWorkspaceActivity';
-import { deriveProjectPath, useWorkspacesQuery } from '../../../../shared/hooks/useWorkspaces';
-import { formatDuration, relativeTime } from '../../../../shared/lib/date';
-import { AGENT_CONTROLLER_ID } from '../chat/services/constants';
-import { isServerFactory, useActiveFactoryContext } from '../workspaces';
-import { FactoryPageShell } from './components/FactoryPageShell';
-import type { FactoryMetrics } from './services/metrics';
-import { BOARD_STAGES, stageLabel, stageOrder } from './stages';
+import { useApiConfig } from '../../../shared/api/config';
+import { useFactoryMetrics } from '../../../shared/hooks/useFactoryMetrics';
+import { useWorkspaceActivity } from '../../../shared/hooks/useWorkspaceActivity';
+import { deriveProjectPath, useWorkspacesQuery } from '../../../shared/hooks/useWorkspaces';
+import { formatDuration, relativeTime } from '../../../shared/lib/date';
+import { AGENT_CONTROLLER_ID } from '../domains/chat/services/constants';
+import { isServerFactory, useActiveFactoryContext } from '../domains/workspaces';
+import { FactoryPageShell } from '../domains/factory/components/FactoryPageShell';
+import type { FactoryMetrics } from '../domains/factory/services/metrics';
+import { BOARD_STAGES, stageLabel, stageOrder } from '../domains/factory/stages';
 
 const WINDOW_OPTIONS = [
   { value: 7, label: '7d' },
